Export pystable as module
