+++ pystable/example.py
@@ -1,7 +1,6 @@
 import os
 import pandas as pd
 import pystable
-import utils
 
 
 def read_helpers(file_name: str):
